# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/texts/{text_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Text",
        "description": "Gets a text by the given id.",
        "operationId": "get_text",
        "parameters": [
          {
            "name": "text_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Text Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextWithQuestionsAndStatistics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete Text",
        "description": "Deletes a text by the given id.",
        "operationId": "delete_text",
        "parameters": [
          {
            "name": "text_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Text Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Text"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/current": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get Current User",
        "description": "Gets the current user's information.",
        "operationId": "get_current_user",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/current/statistics": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get User Statistics",
        "description": "Gets the statistics based on the user's game history.",
        "operationId": "get_user_statistics",
        "parameters": [
          {
            "name": "game_mode",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Mode"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatistics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/current/available_texts": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Get User Available Texts",
        "description": "Gets the texts not read by the user.\nReturns the texts paginated and sorted/filtered by the given parameters.",
        "operationId": "get_user_available_texts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Page Size"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/TextFilter"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Text Filter"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAvailableTexts"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/current/histories": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get Histories",
        "description": "Gets the history of games played by the user.",
        "operationId": "get_histories",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryWithText"
                  },
                  "title": "Response Get Histories"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/current/histories/{history_id}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get History",
        "description": "Gets the history of games played by the user.",
        "operationId": "get_history",
        "parameters": [
          {
            "name": "history_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "History Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryWithQuestions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/register": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Register User",
        "description": "Registers a new user.",
        "operationId": "register_user",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegister"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Login User",
        "description": "Logs in a user. Returns access token and refresh token.",
        "operationId": "login_user",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/logout": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Logout User",
        "description": "Logs out a user. Invalidates the refresh token.\nTODO:\nBlacklist the refresh token.",
        "operationId": "logout_user",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Text": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "fiction": {
            "type": "boolean",
            "title": "Fiction"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "word_count": {
            "type": "integer",
            "title": "Word Count"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content",
          "summary",
          "source",
          "fiction",
          "difficulty",
          "word_count",
          "description",
          "author",
          "image_url",
          "id"
        ],
        "title": "Text"
      },
...
      "TextFilter": {
        "properties": {
          "difficulty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty"
          },
          "include_fiction": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include Fiction",
            "default": true
          },
          "include_nonfiction": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include Nonfiction",
            "default": true
          },
          "only_unplayed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Only Unplayed",
            "default": false
          },
          "keyword": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keyword"
          }
        },
        "type": "object",
        "title": "TextFilter"
      },
...
      "TextWithQuestionsAndStatistics": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "fiction": {
            "type": "boolean",
            "title": "Fiction"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "word_count": {
            "type": "integer",
            "title": "Word Count"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/Question"
            },
            "type": "array",
            "title": "Questions"
          },
          "min_wpm": {
            "type": "integer",
            "title": "Min Wpm"
          },
          "max_wpm": {
            "type": "integer",
            "title": "Max Wpm"
          },
          "average_wpm": {
            "type": "integer",
            "title": "Average Wpm"
          },
          "average_score": {
            "type": "integer",
            "title": "Average Score"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content",
          "summary",
          "source",
          "fiction",
          "difficulty",
          "word_count",
          "description",
          "author",
          "image_url",
          "id",
          "questions",
          "min_wpm",
          "max_wpm",
          "average_wpm",
          "average_score"
        ],
        "title": "TextWithQuestionsAndStatistics"
      },
...
      "User": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "email",
          "role"
        ],
        "title": "User"
      },
      "UserAvailableTexts": {
        "properties": {
          "texts": {
            "items": {
              "$ref": "#/components/schemas/Text"
            },
            "type": "array",
            "title": "Texts"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_texts": {
            "type": "integer",
            "title": "Total Texts"
          }
        },
        "type": "object",
        "required": [
          "texts",
          "page",
          "page_size",
          "total_texts"
        ],
        "title": "UserAvailableTexts"
      },
...
      "UserRegister": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 30,
            "minLength": 3,
            "pattern": "[a-zA-Z0-9_]+",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 30,
            "minLength": 8,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "email",
          "password"
        ],
        "title": "UserRegister"
      },
      "UserStatistics": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "min_wpm": {
            "type": "integer",
            "title": "Min Wpm"
          },
          "max_wpm": {
            "type": "integer",
            "title": "Max Wpm"
          },
          "average_wpm": {
            "type": "integer",
            "title": "Average Wpm"
          },
          "average_wpm_per_day": {
            "items": {
              "$ref": "#/components/schemas/UserStatisticsAverageWpmPerDay"
            },
            "type": "array",
            "title": "Average Wpm Per Day"
          },
          "average_score": {
            "type": "integer",
            "title": "Average Score"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "username",
          "email",
          "min_wpm",
          "max_wpm",
          "average_wpm",
          "average_wpm_per_day",
          "average_score"
        ],
        "title": "UserStatistics"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}